requirements.txt
Django==3.1.7
Pillow=8.1.2
